refactor : ui -> component로 디렉터리명 수정
