+++ feature/running_result/src/main/java/online/partyrun/partyrunapplication/feature/running_result/battle/BattleResultScreen.kt
@@ -43,12 +43,12 @@ import online.partyrun.partyrunapplication.core.designsystem.component.RenderAsy
 import online.partyrun.partyrunapplication.core.model.running_result.ui.BattleResultUiModel
 import online.partyrun.partyrunapplication.core.model.running_result.ui.RunnerStatusUiModel
 import online.partyrun.partyrunapplication.feature.running_result.R
-import online.partyrun.partyrunapplication.feature.running_result.ui.ChartScreen
-import online.partyrun.partyrunapplication.feature.running_result.ui.FixedBottomNavigationSheet
-import online.partyrun.partyrunapplication.feature.running_result.ui.ResultLoadFailedBody
-import online.partyrun.partyrunapplication.feature.running_result.ui.MapWidget
-import online.partyrun.partyrunapplication.feature.running_result.ui.ResultLoadingBody
-import online.partyrun.partyrunapplication.feature.running_result.ui.SummaryInfo
+import online.partyrun.partyrunapplication.feature.running_result.component.ChartScreen
+import online.partyrun.partyrunapplication.feature.running_result.component.FixedBottomNavigationSheet
+import online.partyrun.partyrunapplication.feature.running_result.component.ResultLoadFailedBody
+import online.partyrun.partyrunapplication.feature.running_result.component.MapWidget
+import online.partyrun.partyrunapplication.feature.running_result.component.ResultLoadingBody
+import online.partyrun.partyrunapplication.feature.running_result.component.SummaryInfo
 
 @Composable
 fun BattleResultScreen(
@@ -79,6 +79,7 @@ private fun Content(
     navigateToTopLevel: () -> Unit,
     navigateToBack: () -> Unit
 ) {
+    
     Box(modifier = modifier) {
         when (battleResultUiState) {
             is BattleResultUiState.Loading -> ResultLoadingBody()
