+++ feature/running_result/src/main/java/online/partyrun/partyrunapplication/feature/running_result/component/SummaryInfo.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.feature.running_result.ui
+package online.partyrun.partyrunapplication.feature.running_result.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
